don't set artist tag from album info
setting the artist from developer/publisher can have the following edge cases:
- original file has artist but no album artist. setting album artist to dev (as artist is ignored) will be fine from common sense because the file is part of a game (as a whole) developed by dev
- original file has album artist but no artist. then the artist would be set to dev. this would be incorrect if the original album author is someone else. because this largely implies that album artist also authors the track.

one way to fix this is to skip setting artist if album artist exists in the file. aside of implementation, this can make the value of this tag and the tag fixing process unpredictable.

another way is to sync the artist with album artist. this however is superfluous and not the responsibility of the album info.

the simplest way is to not set the artist tag at all, as having album name and album artist is enough to identify the album.

if the user wants to set the artist tag, they can do so manually by providing extra flags.